EMU Variable: Add support for pre-reserved NV variable store.
Add PcdEmuVariableNvStoreReserved which allows a platform to declare a
memory address for the EMU Variable driver to use for the NV variable
store.  The EMU Variable driver will look to see if the contents of
this memory range appear to be a valid variable store, and if so
the EMU driver will use the variables.

If a platform can preserve a memory range across system resets, this
feature can allow the EMU Variable driver's NV variable store to be
preserved across a system reset.

In the default case this PCD will be set as a fixed PCD with a value
of 0.  In this case this new feature should have minimal impact on
the EMU Variable driver.  (Perhaps a slight increase in code size,
but no functional difference is expected.)

git-svn-id: https://edk2.tianocore.org/svn/edk2/trunk@9240 de2fecce-e211-0410-80a6-f3fac2684e05