DuetPkg: Add DXE APRIORI for 8259 driver
DUET's CpuDxe driver is directly dependent on the 8259.  This
forces the 8259 driver to run before the CPU DXE driver.  This
is required so the 8259 driver will initialize and mask all
8259 interrupts before the CPU DXE driver allows interrupts
to be enabled.

The UefiCpuPkg CPU DXE driver is not dependent on the 8259
directly.  Therefore, in order to use the UefiCpuPkg CPU DXE
driver with DUET, we add the APRIORI file to force the 8259
driver to run early before the CPU DXE driver runs.

git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk@10973 6f19259b-4bc3-4df7-8a09-765794883524