1) Add PcdSerialDetectCable to MdeModulePkg to enable/disable cable detection if hardware flow control is enabled.  The default is to not perform cable detection. 2) Update BaseSerialPortLib16550 to use this new PCD.
This addresses an issue that was introduced on Feb 3, 2011 when all references to Carrier Detect(CD) were removed from the BaseSerialPortLib16550.  When that change was made, a target would block on the first Tx operation if hardware flow control was enabled and no cable or serial application was running on a host. Now the behavior when no cable is connected or no serial app is running on the host is controlled through a PCD.  The default is for the target to continue execution even if there is no cable or serial app running.  If PcdSerialDetectCable is set to TRUE, then the target will block on any transmit that does not have a cable connected or a serial app running in the host.

If hardware flow control is disabled, then this update will have no impact on behavior.


git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk@11301 6f19259b-4bc3-4df7-8a09-765794883524