Use the minimum value for Numeric and the first option value for OneOf as their default value when no specific default value is specified.
HII IFR Numeric and OneOf opcode may not specify their default value in opcode. And, zero may not be valid value for them. To make sure the valid value is set to default value, the minimum value for Numeric and the first option value for OneOf will be chosen as default value. 


git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk@11497 6f19259b-4bc3-4df7-8a09-765794883524