BeagleBoardPkg: Check if the compiler arm-linux-gnueabi-gcc is present in the path
'arm-linux-gnueabi-gcc' is the compiler attached to the ARMLINUXGCC toolchain.
This compiler is available in the PATH environment variable when the Linaro's
toolchain package is installed (Ubuntu distribution).

If if compiler is in the path then select the ARMLINUXGCC toolchain. This change
is to make easier to build UEFI on the BeagleBoard for new starters by hiding the
toolchain.



git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk@12673 6f19259b-4bc3-4df7-8a09-765794883524