MdePkg DxeServicesLib: Search further in GetFileBufferByFilePath
This function supports loading via these interfaces, in order:
* gEfiFirmwareVolume2ProtocolGuid
* gEfiSimpleFileSystemProtocolGuid
* gEfiLoadFile2ProtocolGuid
* gEfiLoadFileProtocolGuid

Previously, if a device path supported any of these interfaces,
it would only check that interface.  (If it failed to load via
that interface, it would not move on and check the others.)

This change causes the other interfaces to still be checked if
the previous ones fail.

Signed-off-by: jljusten
Reviewed-by: mdkinney

git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk@12761 6f19259b-4bc3-4df7-8a09-765794883524